Don't store the locale in a cookie
The locale of tiqr was set in a cookie on a wrong domain (example.com
instead of stepup.example.com), because of this, changed language
settings in other components had no effect on subsequent visits to
tiqr.

Since tiqr no longer has a language switcher, there is no need to save
the locale. It now only reads the stepup_locale cookie and never
changes it.

See https://www.pivotaltracker.com/story/show/156334113