Fix broken registration when enrollment not initialized
PR #40 broke registration if the initial page was loaded before
enrollment is initialized. The actual problem is the tiqr library not
distinguishing between 'session expired' and 'session not yet
started'. To work around this issue, we only interpret the idle status
as 'session expired' when we have seen a status other than idle.