Fix tiqr authentication status
The status of the tiqr authentication was not returned after it
was successful bebcause it would return an expired status because the
check depended on a memcache value which was cleared after success so
the status returned an expired state. The logic order was changed so
first we check for successful authentication and after that we check
if it's expired.