Fix a fatal error in the exception controller
When a unrecoverable error is thrown without a previous exception the
exception controller which displays the error message fails with an
fatal error. This is due to the display handler expecting a previous
exception which is not always the case.

The fatal error is resolved by extending the if condition and checking
for existence of the previous error.

See: https://www.pivotaltracker.com/n/projects/1163646/stories/177016742