Allow clicking of the enrollment QR code
The enrollment link is based on a key that is formed from the user id, a
randomly generated and the request id. These values are turned into a
nonce called the enrollment key. This key is used to form the QR code
and is now also directly used to visit the tiqr registration endpoint
withouth scanning the QR code.

An additional Behat scenario was added, siumlating the qr-click. Given
the design of the Feature, it does not actually click the link, but the
contents of the link are used to set the url, the registration handler
is then visiting. So all in all the correct setting of the anchor href
is verified.

https://www.pivotaltracker.com/story/show/179686777