Refrain from testing window.location changes
This is not in scope of a Jest unit test. The browser should not be
concidered in these kinds of tests. For now I scrapped the test where
this assertion was performed.

An abstraction between the browser and the service could be created if
this proves valuable. That was out of scope for me for now.