Catch ReadWrite errors from Tiqr server library
The read write exceptions are catched in the code calling the library.
The exceptions are explicitly translated to an exception of our own
making. Mainly for decoupling reasons.

The application does not pay further attention to these situations. As
in, it will produce a 500 error when any of these situations arise. No
custom user facing error reports have been prioritized for now.

The biggest win for now is that the application does not continue
execution after a critical point failed previously.