Improve handling of state errors when polling
Prior to this change, when the code expired, the user was shown the error page.
This change shows the user a friendly message and a retry link that serves a fresh qr code.
When a state error occurs, an invalid-request message is sent to the status page, that will then show a 'something went wrong, refresh to try again' message.

See https://www.pivotaltracker.com/n/projects/1163646/stories/188205289