Apply PHPStan checks to the dev folder
Prior to this change none of the code in the dev folder was being scanned by PHPStan, resulting in a lot of errors which do not comply with our coding standards.

This change is needed to improve code quality for our dev code. All existing errors are added to the baseline.