Add the session id derived id to polling requests
Prior to this change, when a polling action from the browser hit the
server, and the server returned a 500, or some other error, it was not
possible to trace which session the browser belonged to.
This change adds a correlation id to the polling mechanism, so that id
can be used to search te logs for the origin of the session.

This should help retrace errors.

See: https://www.pivotaltracker.com/story/show/188205232