Prevent notification authentication MFA fatigue attacks
Prior to this change, if a bad actor could obtain the login credentials
of a user, and the user has push notifications enabled, the bad actor
could trigger a flood of push notifications in the hope the user would
push one.

This change makes it so a push notification will never be sent if the
user did not complete a QR authentication in the browser.

If a QR is scanned, the most recently used notification address is
stored in a secure cookie.
If that address is still the same when a user logs back in, a push
notification will be sent allowing the user to perform MFA.

Resolves https://github.com/OpenConext/Stepup-tiqr/issues/290