Reconfigure dependabot
Prior to this change, dependabot would update all versions. This was not intended, as it causes a lot of pr's, it's not what we want right now.

As per docs, this change sets the `open-pull-requests-limit` to 0. Dependabot will still create pr's for vulnerabilities.

https://docs.github.com/en/code-security/dependabot/dependabot-security-updates/configuring-dependabot-security-updates#overriding-the-default-behavior-with-a-configuration-file

Fixes https://github.com/OpenConext/Stepup-tiqr/issues/336