Fix: Actually apply the trusted device cookie to the browser
Prior to this change, the trusted device cookie was applied to the
response to the tiqr app. Which was unintended, so the user would always
be presented with a qr challenge.
This change puts the cookie on the authentication / registration request
to the browser just before it is redirected back.

Fixes https://github.com/OpenConext/Stepup-tiqr/issues/341